MASSIVE - interceptor is discriminator of aircraft
